feat: add data_types to config and implement populateById functionality in ontPool
